+++ ceammc/ext/doc/docs.cmake
@@ -166,6 +166,7 @@ set(DOC_FILES
     flt.c_lowshelf~
     flt.c_lpf
     flt.c_lpf~
+    flt.c_notch
     flt.c_notch~
     flt.c_pole
     flt.c_peak~