+++ ceammc/CHANGELOG.md
@@ -11,6 +11,7 @@
   - conv.sig2float (with sig->float~ and s->f~ aliases) added
   - flt.c_hpf (with hpf->biquad alias) biquad calculator added
   - flt.c_lpf (with lpf->biquad alias) biquad calculator added
+  - flt.c_notch (with notch->biquad alias) biquad calculator added
   - flt.c_pole (with pole->biquad alias) biquad calculator added
   - loadexpr added (supports functions and datatypes)
   - midi.cc added (enhanced version of ctlin)